Adding STO Female Enlisted
File provided under CC0. This is the enlisted version, and it should be properly sized and the insignia should work properly.

Per the STO guides, enlisted and officers have the same tops with color swaps, so the color swap is all that I did. Left shading shapes and badges the same.

Technically there might be a tiny amount of difference between the greys used for the officer shoulders and the enlisted main body (H22 vs H21); however, without STO open to see the RGB values, and because I don't think it matters visually for this usage, I did not modify that grey.

Also of note: enlisted have different pants, which is not at all relevant for these tokens, but might be worth noting in the event it ever becomes relevant.